+++ src/typesafe/index.ts
@@ -5,4 +5,5 @@ export * from "./errors.js";
 export * from "./questions.js";
 export * from "./responses.js";
 export * from "./retry.js";
+export * from "./cassette.js";
 export * from "./client.js";